Refactor Table1 Component to SchemaTable with Pagination
 - Removed `Table1` component and replaced it with a new `SchemaTable` component.
 - Implemented pagination in `SchemaTable` to handle large datasets efficiently.
 - Updated `ProviderForm.jsx` to use the new `SchemaTable` component.
 - Moved and renamed CSS and asset files from `Table1` to `SchemaTable`.
 - Cleaned up unused code and comments in `ProviderForm.jsx`.
